move RESTFul resources to sep file and update StringIO import for doc creations
